feat: Remove BIOSConfig and add VersionedAttributes (DEL-349) (#7)
This removes the BIOSConfig object type and replaces it with a generic VersionedAttributes. VersionedAttributes match Attributes, except they can't be updated, only created. We also add the ability to filter by versioned attributes when returning hardware same as you can with standard attributes. When searching by versioned attributes we only look at the newest versioned attribute for a given namespace, since that is the current one.