server_component_type: add a Slug field to ServerComponentType and store, return ID as is
Earlier the ServerComponentType.ID was mapped to models.ServerComponentType.Slug
this ends up being unexpected in terms of the data returned, since all
ID's are  UUIDs with the ServerComponentType being the exception.
By adding a Slug field, the data is stored and retrieved transparently,
and this internal remapping does not have to be maintained.

This keeps the server component type db and router models
the same so the Slug doesn't have to be returned as the component type
ID.