server_component: include VersionedAttributes, ComponentTypeSlug fields, struct validator tags
The VersionedAttributes field is to accept firmware version and other possible
component information that is to be stored and looked up.

Add the ComponentTypeSlug field to the struct as well so it can
be converted directly from the models type.

The changes include declaring Gin struct validator tags
to validate input. This pushes the responsibilty to the client
to ensure these fields are declared as expected.