Add server component firmware endpoints (#48)
Adds firmware endpoint to ServerService. This endpoint will be used by two services:

The firmware syncer service which will be responsible for populating the data in the component_firmware_version table.
The inventory collection service which will set the component data during inventory collection (both out of band and in band). The component data model will be updated as part of https://packet.atlassian.net/browse/SE-373 and it should include foreign keys to the component_firmware_version for the current and desired firmware versions.