router/server_component: set component UUID
The INSERT into the Attributes and VersionedAttributes has a check constraint
for server_id (dbSrvComponent.ServerUUID), server_component_id (dbSrvComponent.ID) being NOT NULL,

Generally the INSERT into the server_components table returns a UUID generated by the database
and the dbSrvComponent.ID is set to the returned UUID,

Although, since we're in a transaction here which
INSERTs the component data along with the attributes, versioned attributes in separate statements,
the dbSrvComponent.ID is not set. For this to work, it would require a CTE within which
the returning ID can be assigned to the dbSrvComponent.ID.

For now its easier to just set the UUID here.