chore(migrations): Migrations and components insert fixes (#101)
* db/migrations: idempotent index drop, fixes migrate reset

serverservice migrate reset
2022/07/06 21:12:22 OK    00009_add_firmwares.sql
2022/07/06 21:12:22 OK    00008_remove_deleted_at_index.sql
{"level":"fatal","ts":1657134742.817868,"caller":"cmd/migrate.go:57","msg":"migrate
command
failed","app":"serverservice","version":"dev","command":"reset","error":"failed
to db-down: ERROR 00007_deleted_at_attr.sql

* router/server_component: set component UUID

The INSERT into the Attributes and VersionedAttributes has a check constraint
for server_id (dbSrvComponent.ServerUUID), server_component_id (dbSrvComponent.ID) being NOT NULL,

Generally the INSERT into the server_components table returns a UUID generated by the database
and the dbSrvComponent.ID is set to the returned UUID,

Although, since we're in a transaction here which
INSERTs the component data along with the attributes, versioned attributes in separate statements,
the dbSrvComponent.ID is not set. For this to work, it would require a CTE within which
the returning ID can be assigned to the dbSrvComponent.ID.

For now its easier to just set the UUID here.

* router/server_component: fixes component ID and INSERTs component Attributes.

The component ID is now referenced from the DB model object where its set,
earlier this field was just empty and the INSERT would fail.

The component Attributes were not being inserted earlier.

* router/server_component: purge Gin validation for vendor, model, serial fields

Components can have these fields missing and so they should not be
required.

* db/migrations: fixes component attributes check contraint

The check_server_id_server_component_id check constraint expects
a return value of 1::INT8 while, when both the server_id and component_id is set the check
constraint returns 2::INT8

select ((a IS NOT NULL)::INT8 + (b IS NOT NULL)::INT8 = 2::INT8) from
(values('fc167440-18d3-4455-b5ee-1c8e347b3f36'::uuid,
'c4eb3a11-e665-4f9e-b4ae-1984e5c652b4')) s(a, b);

* db/migrations: remove index idx_server_namespace from attributes, versioned_attributes

This index prevents multiple component attributes that reference the same server ID
from being inserted. Purging this index until a better work around is determined.

hollow client received a server error - response code: 500, message: datastore
error, details: failed to insert into foreign table: models: unable to insert
into attributes: pq: duplicate key value violates unique constraint "idx_server_namespace"

* docker: add dockerignore to reduce build context size

* router/server_components: revert check constraint change

Based on feedback in the PR, the constraint is as expected -
either one of the ServerUUID OR the ComponentID should be set for a row in attributes, versioned_attributes.

* api/server_component: set Serial, Component Slug as required fields

This is to add a unique index so we don't end up with duplicate
component information. In the case where a component as no serial,
the caller is expected to set an incrementing value. Its upto the caller
to ensure the incrementing value is consistent over multiple data
collections.

* db/migrations: server_components - Add a composite index to prevent duplicates

This adds a unique index on the server_id, serial, server_component_type_id columns
to ensure the same component will not be added multiple times.

* router/server_components: purge versioned_attributes group by query modifier

* api/pagination: remove duplicate Attributes query modifier

* revert: remove index idx_server_namespace from attributes, versioned_attributes

This adds back the idx_server_namespace index.

The change being reverted was made on the assumption that both the ServerID
and the ComponentID are to be declared for each component attribute.

Based on the discussion in the PR https://github.com/metal-toolbox/hollow-serverservice/pull/101#discussion_r915225582
either the ServerID or the Server ComponentID should be set.

reverts commit:

commit 988350ab683150a42c1393f39aa4f2a75f7d4be2
Author: Joel Rebello <jrebello@packet.com>
Date:   Wed Jul 6 21:56:51 2022 +0200

    db/migrations: remove index idx_server_namespace from attributes, versioned_attributes

    This index prevents multiple component attributes that reference the same server ID
    from being inserted. Purging this index until a better work around is determined.

    hollow client received a server error - response code: 500, message: datastore
    error, details: failed to insert into foreign table: models: unable to insert
    into attributes: pq: duplicate key value violates unique constraint "idx_server_namespace"

* api/router_server_components: component results to include just the latest VersionedAttribute

This change limits the returned VersionedAttributes to the latest
so we're not including all of the versioned attributes.