loadServerFromParams(): expect caller to set JSON response
In order to reduce ambiguity in who writes the response body,
this method now just returns an error if any. and leaves it
to the caller to write the response based on the error.

Earlier this method was writing a JSON response to the body, which led to multiple
situations where the JSON response was written again by the caller. And
so invalid JSON being sent in the JSON body.