api/router_server_components: upsert component attribute
This simplifies work for clients when updating component information,
by performing an update when a certain attribute is present OR and
insert when its not already present.

Without this change, clients get returned a 400 - 'no sql rows in result',
when a component attribute isn't present, which is then cumbersome to deal with on the clientside.

Note:

The update, insert could have been swapped with the sqlboil Upsert() method,
although since the attributes table contains a partial index - "WHERE server_component_id is not null"
and the sqlboil Upsert() method has no way of specifying query mods,
the current Upsert() method does not work for this case (the matching row for update is not found in the row scan).

https://github.com/volatiletech/sqlboiler/issues/856