db/migrations: Adds tables for server condtions, condition types, condition status objects
The server_conditions table holds a record to a 'condition' object that is
finalized on a server by a controller (reconciler). Its expected that
a controller updates this table as it finalizes the work defined in the
condition.

The producer (client) creates the condition with the `parameters` field containing
a JSON object. The consumer (controller) updates the condition in the `status_output` field
containing a JSON object. The format of these JSON object fields is left to the producer and consumer of the data.

The server_condition_types holds the possible condition 'types' that can
be set in the server_conditions table. A few condition types are created
as part of the migration - `firmwareUpdate`, `inventoryOutofband` these
are expected to be used by the Alloy and Flasher controllers.

The server_condition_status_types hold the possible condition statuses
that can be set in the server_conditions table. A few condition status
types are created as well.