api/v1/attribute_list_params: implement support for Attribute Operators
Querying multiple Attributes would always use the 'AND' operator in the
SQL query, this is limiting and required to support the 'OR' operator
when required by the client.

This change extends the AttributeListParams to support an Attribute Operator - 'and' 'or'
to enable the client to define how multiple Attributes should be SQL queried.

This change is backwards compatible and defaults to 'and' which is the previous
behaviour.

The client can query multiple attributes and specify which of them should looked up
with an OR, by specifying a final '~' separated value - 'or'.

"attr=hollow.versioned~a.name~like~nemo&attr=hollow.versioned~a.name~like~bluefin~or"