dbtools/testtools: cleanDB() to log errors if any fixture fails to delete as part of tests
The cleanDB() method is invoked multiple times for each test,
if a table cannot be deleted because of a constraint or some other
cause, this method will now log the error as part of the test.

This helps debug cause for failing tests quicker - in the case where the
DB tables were not cleaned up by cleanDB().